Fix for CSCva28889
Fix is to delete the cfg file in the disk. Bug description is as follows:
In the event that user changes from applied policy to another policy with
different rules, the ASA configuration shows both older rules and new rules.
steps:
openstack is configured for FW1 with policy1 and  x number of rules.
Delete the current firewall, create new rules and a new policy (policy2).
Add new rules to the new policy.
create a FW and apply the new policy to the FW.
ASA configuration shows both old and new rules.