Fix to verify the response from lldpad for ass
When VDP (lldpad) sends a response to enabler agent for a associate
message, enabler doesn't check if the UUID of port in the request
matches with that of the response from VDP. Even though, it always matches,
there were few cases, where the response does not match. This fix
takes care of that. DDTS is CSCvb62407.