Implement OVS flow checker for data traffic
For SAF/DFA/PF, the enabler programs the OVS flows in br-int and br-ethd.
Those OVS bridges are generally owned by Openstack and there's always a
possibility that flows can get cleared for some conditions. When OVS Neutron
agent is restarted the OF rules are cleared. When this happens, the flows
in br-int gets cleared for sure. Regarding br-ethd:
a. In RHEL-OSP setups, we have seen  the flows getting cleared for br-ethd
as well. In this case, the LLDP packets from the switch will not make it to
the LLDPAD in server because the rules for LLDP/VDP from br-ethd also gets
cleared. So, neighbor information for LLDPAD will be lost as a result of
which uplink from OVS bridge gets removed. Generally, VSI associate from OVS
bridge will also return failure due to which enabler will remove the VDP
VLAN. Then, regular uplink detection kicks in and both lldp/vdp flows as
well as data flows are put back eventually.
b. The br-ethd flows do not get cleared, but the br-int flows gets cleare. In
this case, the enabler will not put the flows back since the LLDP neighbor for
the server is not lost and the VDP associate for all VSI's from LLDPAD returns
the right VLAN. So, the enabler has no way of knowing the flows are deleted.

So, for the above case (b) and for any unknown corner cases, it was decided to
implement a periodic checker that will ensure that the flows are put back, if
deleted.