Server not sending the Bulk VM list after restarts
This case was observed, when upgrade was being tested
in a OSP7 HA setup. Here, enabler server got restarted
thrice (which eventually got fixed) and even the enabler
agent got restarted.

The flow of events, when an agent starts up is:
a. The agent will send a heartbeat to the server
b. The server will establish an RPC with the agent after
receiving the heartbeat.
c. The agent will save the uplink into the server DB.
d. The server will send all the list of VM's running on
that compute node, to the agent after receiving the uplink.
It uses the RPC information created in step no.b, to send
that information.

Now, during upgrade what happened was the agent performed
step (a) and the server running in controller received it.
The server restarted and the agent performed step (c) and
since the server did not receive any heartbeat (yet) from the
agent, it will not be able to bulk send the list of VM's on
that agent. The server does not have a retry mechanism for sending
the list of VM's to the agent as it was designed such that it was
agent's job to save the uplink as a resut of which the server sent
all the VM list to the agent.
In keeping with the design, this DDTS enhances the agent code to:
a. Have a flag to signify if the bulk VM list is obtained from the server
b. If it is not obtained, it will do a save_uplink during periodic
uplink detection which will trigger the server to send the list of VM's.
The same issue also happens during request uplink info, which is also
taken care of.