Propagate VM auto-config failure reason to server
Currently, when auto config fails for a VM, the reason is not propagated
back to the server. One needs to look into the agent's log files to figure
out the reason. This commit takes care of propagating the failure reason back
to the server, which can be queried from the DFA server using the CLI.
This information is not stored in the persistent DB. It's maintained as a
cache. The agent sends the failure reason periodically or if there's a
change in the failure reason.